chore: resolve TODO for `ServiceAccount`-only rotation
Initially, PKS maintainers decided to rotate only `ServiceAccount`s
since we have a `Secret` that represents `API_KEY` for PoolC API server
in the same namespace. However, it turns out that separating namespaces
for `ServiceAccount`s and `CronJob` is a way simpler approach.

In the current settings, the `CronJob` and manifests needed by it will
live in the `poolc-system` namespace, whereas `ServiceAccount`s that
identifies PoolC users will live in the `poolc-users` namespace.