perf(ci): stop building image for arm64
The initial purpose of applying multi-platform build is to enable
testing on local environments such as Apple Silicon MacBooks. However,
it turns out that tests can be done on our servers quite conveniently.
As a result, the delay caused by multi-platform build outweighs the
benefit it provides.

If multi-platform build is re-enabled in the future for some reasons,
the following resources might be helpful:
- https://docs.github.com/en/packages/working-with-a-github-packages-registry/working-with-the-container-registry#adding-a-description-to-multi-arch-images
- https://docs.docker.com/build/ci/github-actions/multi-platform/#distribute-build-across-multiple-runners