feat(log): specify 'PoolC API server' in log and error messages
Previously, log messages and errors used the generic term 'API', which
was ambiguous because the credentials updater interacts with both the
kube-apiserver and the PoolC API server.

Update messages to explicitly name the 'PoolC API server' to make logs
and errors clearer and avoid potential confusion when debugging
interactions with different servers.